+++ sentry/src/test/java/io/sentry/SentryOptionsTest.kt
@@ -248,6 +248,6 @@ class SentryOptionsTest {
 
     @Test
     fun `when options are initialized, maxAttachmentSize is 20`() {
-        assertEquals(20 * 1024 * 1024, SentryOptions().maxAttachmentSize)
+        assertEquals((20 * 1024 * 1024).toLong(), SentryOptions().maxAttachmentSize)
     }
 }
